% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lod.R
\name{lod}
\alias{lod}
\title{Single-point LOD score}
\usage{
lod(
  x,
  aff,
  model,
  markers = NULL,
  rho = 0,
  maxOnly = NA,
  loop_breakers = NULL,
  peelOrder = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{A \code{ped} object.}

\item{aff}{A vector of length \code{pedsize(x)}, containing the affection status
of each pedigree member (2 = affected; 1 = unaffected; 0 = unknown).
Alternatively, a list of vectors named \code{affected} and (optionally)
\code{unaffected}.}

\item{model}{A \code{disModel} object, typically created with \code{\link[=diseaseModel]{diseaseModel()}}.}

\item{markers}{A vector of marker names or indices referring to markers
attached to \code{x}. By default all markers are included.}

\item{rho}{A number between 0 and 0.5 (inclusive); the hypothesised
recombination ratio between the marker and the disease locus.}

\item{maxOnly}{a logical indicating whether only the maximum LOD score should
be returned. By default this is always done if the number of markers is 1.}

\item{loop_breakers}{A vector of ID labels indicating loop breakers. (Only
relevant for inbred pedigrees.)}

\item{peelOrder}{For internal use.}

\item{verbose}{a logical: verbose output or not.}
}
\value{
If the number of markers is 1, or if \code{maxOnly = TRUE}, a single
number is returned.

Otherwise a \code{linkres} object, which is basically a data frame with columns
\code{CHROM}, \code{MARKER}, \code{MB} and \code{LOD}.
}
\description{
Calculates the single-point LOD scores of a pedigree for the specified
markers. The recombination ratio between the disease and marker loci can be
either fixed or optimized.
}
\details{
The LOD score of a marker is defined as \deqn{LOD(\rho) = \log\link{10}
\frac{L(\rho)}{L(0.5)}}{\% LOD(\rho) = log10 L(\rho)/L(0.5)} where
\eqn{L(\rho)} denotes the likelihood of the observed marker genotypes given a
recombination ratio \eqn{\rho} between the marker and the disease locus.
}
\examples{

x = nuclearPed(2)
x = setMarkers(x, marker(x))
genotype(x, 1, c(1,2:4)) = 1:2
genotype(x, 1, 2) = 1

aff = c(2,1,2,2)
model = diseaseModel(model = "AD")

lod(x, aff, model)

}
\seealso{
\link{linkres}, \code{\link[=merlinLod]{merlinLod()}}, \code{\link[=diseaseModel]{diseaseModel()}}, \code{\link[=lodPeaks]{lodPeaks()}}
}
\author{
Magnus Dehli Vigeland
}
