% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramlink2-package.R
\docType{package}
\name{paramlink2-package}
\alias{paramlink2}
\alias{paramlink2-package}
\title{paramlink2: Parametric Linkage Analysis}
\description{
Parametric linkage analysis of monogenic traits in medical pedigrees. Features include singlepoint analysis, multipoint analysis via 'MERLIN' (Abecasis et al. (2002) \doi{10.1038/ng786}), visualisation of log of the odds (LOD) scores and summaries of linkage peaks. Disease models may be specified to accommodate phenocopies, reduced penetrance and liability classes. 'paramlink2' is part of the 'pedsuite' package ecosystem, presented in 'Pedigree Analysis in R' (Vigeland, 2021, ISBN:9780128244302).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/paramlink2}
  \item Report bugs at \url{https://github.com/magnusdv/paramlink2/issues}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

}
\keyword{internal}
