% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramlink2-package.R
\docType{package}
\name{paramlink2}
\alias{paramlink2}
\title{paramlink2: Parametric linkage analysis}
\description{
Parametric linkage analysis of monogenic traits in medical
pedigrees. Features include singlepoint analysis, multipoint analysis
(through a wrapper to the 'MERLIN' software), visualisation of LOD
scores and summaries of linkage peaks.
}
