% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramlink2-package.R
\docType{package}
\name{paramlink2}
\alias{paramlink2}
\title{paramlink2: Parametric linkage analysis}
\description{
Parametric linkage analysis of monogenic traits in medical pedigrees.
Features include singlepoint analysis, multipoint analysis via 'MERLIN'
(Abecasis et al. (2002) \doi{10.1038/ng786}), visualisation of log of the
odds (LOD) scores and summaries of linkage peaks. Disease models may be
specified to  accommodate phenocopies, reduced penetrance and liability
classes. 'paramlink2' is part of the 'ped suite' package ecosystem, presented
in 'Pedigree Analysis in R' (Vigeland, 2021, ISBN:9780128244302).
}
